added difficulties, fixed board algorhitm
so initially I write the board generation and clearing in a opposite way
than intended. So when I tested out non square boards, only the squared
part of the board would be cleared (because it couldn't get to the
part that didn't exist, which program thought it was in the opposite of
the correct way.
I doubt that I explained what was the problem corectly, but just know
that it's fixed now